Split cluster-details.tsx
